caselaw_service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
datasets==2.19.0
sentence-transformers==2.6.1
python-dotenv==1.0.0
httpx==0.27.0
pytest==8.1.1
pytest-asyncio==0.23.6
python-jose==3.3.0
